💚 ci: install unxts.parametric in the workspace extra; relocate its tests
The `unxt` test matrix failed to collect because core tests imported
`unxts.parametric`, which the job did not install: the
`[project.optional-dependencies].workspace` extra listed the other
workspace packages but was missing `unxts.parametric`. Add it (and relock).

Also make core self-contained regardless: relocate the parametric-dependent
tests out of core `tests/` into the parametric package
(test_quantity_staticvalue_eq, test_static_quantity_reparent, and the
pow-with-PQ-exponent case → test_pow_exponent), and collect-ignore the
parametric package docs when `unxts.parametric` is not installed (matching
the interop-package convention).
