+++ packages/unxts.parametric/tests/test_quantity_staticvalue_eq.py
